Refs #6007. Add rmw_compare_gids_equal.c file.
