+++ rmw_microxrcedds_c/src/rmw_microxrcedds.c
@@ -494,25 +494,6 @@ rmw_ret_t rmw_get_gid_for_publisher(const rmw_publisher_t * publisher, rmw_gid_t
   return RMW_RET_OK;
 }
 
-rmw_ret_t rmw_compare_gids_equal(const rmw_gid_t * gid1, const rmw_gid_t * gid2, bool * result)
-{
-  // Check
-  RMW_CHECK_ARGUMENT_FOR_NULL(gid1, RMW_RET_INVALID_ARGUMENT);
-  RMW_CHECK_ARGUMENT_FOR_NULL(gid2, RMW_RET_INVALID_ARGUMENT);
-  if (gid1->implementation_identifier != rmw_get_implementation_identifier()) {
-    RMW_SET_ERROR_MSG("publisher handle not from this implementation");
-    return RMW_RET_INCORRECT_RMW_IMPLEMENTATION;
-  } else if (gid2->implementation_identifier != rmw_get_implementation_identifier()) {
-    RMW_SET_ERROR_MSG("publisher handle not from this implementation");
-    return RMW_RET_INCORRECT_RMW_IMPLEMENTATION;
-  }
-
-  *result =
-    memcmp(gid1->data, gid2->data, sizeof(rmw_gid_t)) == 0;
-
-  return RMW_RET_OK;
-}
-
 rmw_ret_t rmw_service_server_is_available(
   const rmw_node_t * node, const rmw_client_t * client,
   bool * is_available)
